Simplify canonicalUrl API to only require page path
Move site.json import and URL validation into the utility module so
callers no longer need to pass the site URL - they just provide the
page path and the canonical URL is computed automatically.
